completes SAT collision and translates origin
This commit debugs the SAT collision so that it detects
when the objects collide. More work needs to be done to
actually repel objects when they collide. ALso, this
commit changes the code so that the origin is translated
from the top left corner to the middle of the pane.