removes debugging code and improves collision
This commit removes debugging code from testing collision
detection. It also improves the collision itself. Now,
when the main character is turned into the target, it slides
along the side of the character and is pushed back to prevent
overlap. Also, the logic used to move the main character
forwards until right up against the target's bounding box was
refined. Finally, the main character image was switched to a
tank.