fixes collision logic and removes code for debugging
This commit fixes some of the logic for collision by compensating
for the scenario where the main character isn't moving but collides
with an object because of a change in angle. This change allows for
the rotation and instead pushes the main character backwards until
the two game elements are no longer colliding. They sometimes phase
through each other if one hits the other almost parallel, but for
all other scenarios the collision works. In order to clean up the
code, some debugging code was removed and some comments were added.