fixes bug with collision when multiple targets are present
This commit changes the code to look at all the elements
on the map when trying to detect a collision. Code was
added to debug this and make it work. Also, it will
need to be optimized to look at only a list of the
game elements within a certain range of the main
character, instead of all the elements on the map.