adds changes suggested by javagl
This commit takes care of a few things. First, it
moves the control logic out of the MainPanel and into
its own class. Then, the Point class is removed. Instead,
the prebuilt Point2D.Double class is used. Some values may
be changed to Float in the future to save space, but everything
remains as double for now. Also, the class Map was renamed to
GameMap to avoid confusing it with the prebuilt java.util.Map
class. Finally, the Utility class was created to contain
usefull methods like vectorDotProduct, getVectorMagnitude, etc.