fixes bug related to sprite movement
This commit fixes a bug related to the movement of
sprites and modifies the GameMap class with changes
suggested by javagl. Also, code to test the changes
was added. This meant that when 'r' is pressed, a
new sprite is added to the map with a certain speed
and a certain rotational speed.