improves turning logic and adds comments
This commit completely changes the turning logic to
account for additions in the future. Instead of instantly changing
the turning speed to a certain value while turning and to zero while
stopped, the Sprites have a rotational velocity that is added or subracted
from while turning or stopping. This allows modifications in the future
where the Sprite's base rotational velocity can be changed by an external
factor (like standing on a rotating object). Instead of overriding that
rotation when the sprite is told to stop or start turning, the new logic
adds or subtracts from the rotation relative to how fast it is rotating
already.