makes Hitbox logic based off of inheritance
This commit adds a RectangleHitbox class and a
CowHitbox class, both of which inherit the now abstract
class HitBox. The difference bewteen the two classes is
the computeCollisionPoints method, which changes depending
on the shape of the Sprite. Some debug logic was also added
to switch the shape of the MainCharacter bewteen a cow and
a tank to test the transition bewteen hitboxes.